CSUS colors and general template
Workin' on making it pretty! In progress, a couple things are off still
but once we get the functionality added to the other pages (database
entry for tours, sites, surveys along with pulling data for the
statistics page) it should be quick. Paste your page modifications into
the "Content_Area" div